[DE-401] Java Driver v7 migration (#268)
* Java Driver v7 migration

* converter migration

* fixed handling of null or missing fields

* fixed byte fields handling

* fixed big decimal converter

* fixed custom mapping tests

* fixed ArangoTemplateTest

* fixed AQL fullCount stats deserialization

* removed vpack usages

* rm milestone repos

* upd enforcer version

* removed deprecated usages of ObjectNode#put()

+++ src/test/java/com/arangodb/springframework/repository/query/ArangoAqlQueryTest.java
@@ -64,7 +64,7 @@ public class ArangoAqlQueryTest extends AbstractArangoRepositoryTest {
 		assertThat(retrieved, hasEntry("_rev", john.getRev()));
 		assertThat(retrieved, hasEntry("customer-name", john.getName()));
 		assertThat(retrieved, hasEntry("surname", john.getSurname()));
-		assertThat(retrieved, hasEntry("age", (long) john.getAge()));
+		assertThat(retrieved, hasEntry("age", john.getAge()));
 		assertThat(retrieved, hasEntry("alive", john.isAlive()));
 	}
 
@@ -76,7 +76,7 @@ public class ArangoAqlQueryTest extends AbstractArangoRepositoryTest {
 		assertThat(retrieved.getRevision(), is(john.getRev()));
 		assertThat(retrieved.getAttribute("customer-name"), is(john.getName()));
 		assertThat(retrieved.getAttribute("surname"), is(john.getSurname()));
-		assertThat(retrieved.getAttribute("age"), is((long) john.getAge()));
+		assertThat(retrieved.getAttribute("age"), is(john.getAge()));
 		assertThat(retrieved.getAttribute("alive"), is(john.isAlive()));
 	}
